# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/dashboard:
    get:
      operationId: getDashboardMetrics
      tags: [dashboard]
      summary: CRM dashboard metrics
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/portal/dashboard:
    get:
      operationId: portalDashboard
      tags: [portal]
      summary: Client portal executive dashboard
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalDashboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/portal/organisation:
    get:
      operationId: portalOrganisation
      tags: [portal]
      summary: Organization workspace profile
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalOrganisationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/portal/users:
    get:
      operationId: portalUsers
      tags: [portal]
      summary: Organization members
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalMemberListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/portal/projects:
    get:
      operationId: portalProjects
      tags: [portal]
      summary: Organization projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalProjectListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/portal/documents:
    get:
      operationId: portalDocuments
      tags: [portal]
      summary: Organization documents
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalDocumentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/portal/support:
    get:
      operationId: portalSupportTickets
      tags: [portal]
      summary: Support tickets
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSupportTicketListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      operationId: createPortalSupportTicket
      tags: [portal]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortalSupportTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSupportTicket"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/portal/activity:
    get:
      operationId: portalActivity
      tags: [portal]
      summary: Unified activity timeline
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalActivityListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/audits/dashboard:
    get:
      operationId: auditDashboard
      tags: [audits]
      summary: Audit executive dashboard
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/audits:
    get:
      operationId: listAudits
      tags: [audits]
      summary: List compliance audits
      responses:
        "200":
          description: Audit list
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAudit
      tags: [audits]
      summary: Create audit plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Audit created
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/audits/{auditId}:
    get:
      operationId: getAudit
      tags: [audits]
      summary: Audit detail
      parameters:
        - name: auditId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audit detail
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/audits/calendar:
    get:
      operationId: auditCalendar
      tags: [audits]
      summary: Audit calendar events
      responses:
        "200":
          description: Calendar events
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/compliance/dashboard:
    get:
      operationId: complianceDashboard
      tags: [compliance]
      summary: Compliance dashboard widgets
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/compliance/analytics/executive:
    get:
      operationId: complianceAnalyticsExecutive
      tags: [compliance]
      summary: Executive compliance analytics dashboard
      responses:
        "200":
          description: Executive dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceExecutiveDashboard"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/compliance/analytics/regulatory:
    get:
      operationId: complianceAnalyticsRegulatory
      tags: [compliance]
      summary: Regulatory monitoring and alerts
      responses:
        "200":
          description: Regulatory intelligence
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/compliance/analytics/alerts:
    get:
      operationId: complianceAnalyticsAlerts
      tags: [compliance]
      summary: Regulatory alert inbox
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/inspections:
    get:
      operationId: listInspections
      tags: [inspections]
      summary: List inspections
      responses:
        "200":
          description: Inspection list
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createInspection
      tags: [inspections]
      summary: Create inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/inspections/dashboard:
    get:
      operationId: inspectionDashboard
      tags: [inspections]
      summary: Inspection dashboard
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/capa:
    get:
      operationId: listCapa
      tags: [capa]
      summary: List CAPA records
      responses:
        "200":
          description: CAPA list
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createCapa
      tags: [capa]
      summary: Create CAPA record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, capaType]
              properties:
                title:
                  type: string
                capaType:
                  type: string
                  enum: [corrective, preventive]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/capa/dashboard:
    get:
      operationId: capaDashboard
      tags: [capa]
      summary: CAPA dashboard metrics
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/capa/{capaId}:
    get:
      operationId: getCapa
      tags: [capa]
      summary: Get CAPA detail
      parameters:
        - name: capaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CAPA detail
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateCapa
      tags: [capa]
      summary: Update CAPA record
      parameters:
        - name: capaId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/capa:
    get:
      operationId: portalListCapa
      tags: [portal, capa]
      summary: List CAPA records for client portal
      responses:
        "200":
          description: Portal CAPA list
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/capa/dashboard:
    get:
      operationId: portalCapaDashboard
      tags: [portal, capa]
      summary: CAPA dashboard for client portal
      responses:
        "200":
          description: Portal CAPA dashboard
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/risk-assessments:
    get:
      operationId: listRiskAssessments
      tags: [risk-assessments]
      summary: List risk assessments
      responses:
        "200":
          description: Risk assessment list
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createRiskAssessment
      tags: [risk-assessments]
      summary: Create risk assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
      responses:
        "201":
          description: Created
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/risk-assessments/dashboard:
    get:
      operationId: riskAssessmentDashboard
      tags: [risk-assessments]
      summary: Risk dashboard metrics
      responses:
        "200":
          description: Dashboard
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/risk-assessments/heatmap:
    get:
      operationId: riskHeatMap
      tags: [risk-assessments]
      summary: Risk heat map data
      responses:
        "200":
          description: Heat map
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/risk-assessments:
    get:
      operationId: portalListRiskAssessments
      tags: [portal, risk-assessments]
      summary: List risk assessments for client portal
      responses:
        "200":
          description: Portal risk list
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/risk-assessments/dashboard:
    get:
      operationId: portalRiskDashboard
      tags: [portal, risk-assessments]
      summary: Risk dashboard for client portal
      responses:
        "200":
          description: Portal risk dashboard
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/compliance:
    get:
      operationId: portalCompliance
      tags: [portal, compliance]
      summary: Client compliance workspace aggregate
      responses:
        "200":
          description: Portal compliance workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalComplianceResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/compliance/analytics/executive:
    get:
      operationId: portalComplianceAnalyticsExecutive
      tags: [portal, compliance]
      summary: Client executive compliance analytics
      responses:
        "200":
          description: Executive analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceExecutiveDashboard"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/compliance/analytics/regulatory:
    get:
      operationId: portalComplianceAnalyticsRegulatory
      tags: [portal, compliance]
      summary: Client regulatory monitoring view
      responses:
        "200":
          description: Regulatory analytics
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/compliance/analytics/alerts:
    get:
      operationId: portalComplianceAnalyticsAlerts
      tags: [portal, compliance]
      summary: Client open regulatory alerts
      responses:
        "200":
          description: Open alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegulatoryAlertListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portal/capa/{capaId}:
    get:
      operationId: portalGetCapa
      tags: [portal, capa]
      summary: CAPA detail for client portal
      parameters:
        - name: capaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CAPA detail
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
...
    Unauthorized:
      description: Authentication is required, or the session has expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: >
        Authenticated, but lacking the required permission or tenant scope.
        Retrying with the same credentials will not succeed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Too many requests. Retry after the window resets.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: >
        An unexpected fault. The message is generic by design; quote
        `error.requestId` when reporting it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: The service is temporarily unable to handle the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    DashboardMetrics:
      type: object
      properties:
        newLeads:
          type: number
        qualified:
          type: number
        won:
          type: number
        lost:
          type: number
        openLeads:
          type: number
        averageResponseMinutes:
          type: number
          nullable: true
      required:
        - newLeads
        - qualified
        - won
        - lost
        - openLeads
...
    PortalOrganisationResponse:
      type: object
      additionalProperties: true
...